pyproject.toml
[project]
name = "churn-prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "comet-ml==3.40.0",
    "fastapi>=0.115.12",
    "fire>=0.7.0",
    "lightgbm>=4.6.0",
    "numpy>=2.2.4",
    "optuna>=4.2.1",
    "pandas>=2.2.3",
    "pyarrow>=19.0.1",
    "pydantic>=2.11.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "streamlit>=1.44.0",
    "uvicorn>=0.34.0",
    "xgboost>=3.0.0",
]
